refactor: 라우팅 URL 파싱 로직 수정
